pyproject.toml
[project]
name = "Inductor"
version = "0.0.1"
description = "Code Explorer for HLASM"
authors = [
    { name = "Avishek Sen Gupta", email = "foo@bar.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "langgraph",
    "python-dotenv",
    "mcp",
    "mcp[cli]",
    "uv",
    "langchain-mcp-adapters",
    "langchain-anthropic",
    "openai",
    "mypy",
    "boto3",
    "awscli",
    "langchain",
    "anthropic",
    "langchain-aws",
    "neo4j",
    "dataclasses_json"
]

[tool.poetry]
name = "Inductor"  # This should match your expected package name
version = "0.0.1"
description = "Code Explorer for HLASM"
packages = [
    {include = "src"}  # Make sure this points to your source directory
]

[tool.poetry.dependencies]
python = "^3.13"

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1", "pytest>=7.0.0"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"
